requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
openai
python-multipart
pydantic
pydantic-settings
numpy
